entry_task/requirements.txt
Django==1.11.29
gunicorn==19.10.0
django-mysql==2.2.0
mysqlclient==1.3.14
pyjwt==1.7.1
requests==2.24.0
gevent==20.6.2
mysql-connector-python==8.0.21